requirements.txt
SQLAlchemy~=2.0.22
werkzeug~=3.0.0
Pillow~=10.1.0
flask~=3.0.0
python-dotenv~=1.0.0
asgiref~=3.7.2